requirements.txt
asgiref==3.5.0
attrs==21.4.0
Django==4.0.3
django-environ==0.4.5
djangorestframework==3.13.1
freezegun==1.2.1
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pyparsing==3.0.7
pytest==6.2.5
pytest-django==4.5.2
pytest-freezegun==0.4.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
